requirements.txt
discord.py
firebase-admin
Flask==2.3.2